refactor: コンポーネントのリファクタリングとフォルダー構造の見直し
