+++ packages/zenn-embed-components/src/Icons/GithubStarIcon.tsx
