+++ packages/zenn-embed-components/src/components/link-card/EmbedBaseLinkCard.tsx
@@ -1,6 +1,6 @@
 import { css } from '@emotion/react';
 import { useState } from 'react';
-import { GithubIcon } from '../Icons/GithubIcon';
+import { GithubIcon } from '../../Icons/GithubIcon';
 import { LinkData } from './types';
 
 export interface EmbedBaseLinkCardProps extends LinkData {}
