+++ packages/zenn-embed-components/src/index.ts
@@ -1,1 +1,5 @@
-export { EmbedGitHub } from './components/github';
+export * from './components/github';
+export * from './components/gist';
+export * from './components/link-card';
+export * from './components/mermaid';
+export * from './components/tweet';
